+++ c/__float128.c
@@ -1,3 +1,12 @@
+/**
+ * __float128 is available on i386, x86_64, IA-64, LoongArch and hppa HP-UX,
+ * as well as on PowerPC GNU/Linux targets that enable the vector scalar (VSX)
+ * instruction set. __float128 supports the 128-bit floating type. On i386,
+ * x86_64, PowerPC, LoongArch and IA-64, other than HP-UX, __float128 is an
+ * alias for _Float128. On hppa and IA-64 HP-UX, __float128 is an alias for
+ * long double.
+ */
+#include <assert.h>
 #include <stdio.h>
 #include <stdint.h>
 
@@ -17,9 +26,13 @@
 
 int main(void)
 {
-	__float128 f128 = 3.14;
+	assert(sizeof(__float128) == 16 && "size of __float128 is not equal to 16");
 
-	(void)f128;
+	__float128 f128 = 3.14;
+	_Float128 F128 = 3.14;
+
+	printf("size of __float128 %ld\n", sizeof(f128));
+	printf("size of _Float128 %ld\n", sizeof(F128));
 
 	return 0;
 }